fix: Implement BLE Auto-Reconnect and Repeat Pairing for Windows
- Added BLE_GAP_EVENT_REPEAT_PAIRING handler to auto-delete stale bonds when Windows removes the device and tries to re-pair, preventing pairing deadlocks.
- Injected Device Name into BLE Scan Response Data so Windows 11 can identify the advertising gamepad and trigger Auto-Reconnect without user intervention.

+++ components/ble_hid/include/ble_hid.h
@@ -23,9 +23,9 @@ void ble_hid_deinit(void);
  * @param roll     1000-2000
  * @param p1       1000-2000 (Potentiometer 1)
  * @param p2       1000-2000 (Potentiometer 2)
- * @param buttons  Bitmask of 8 buttons
+ * @param buttons  Bitmask of 16 buttons
  */
-void ble_hid_send_report(uint16_t throttle, uint16_t yaw, uint16_t pitch, uint16_t roll, uint16_t p1, uint16_t p2, uint8_t buttons);
+void ble_hid_send_report(uint16_t throttle, uint16_t yaw, uint16_t pitch, uint16_t roll, uint16_t p1, uint16_t p2, uint16_t buttons);
 
 /**
  * @brief Check if BLE is connected to a host
